Corrected imports in example
Due to moving the files the imports need to be corrected as well.
With example_plot_UEs.py in its own example directory it can no longer
directly import the relevant function for Unitary Event Analysis.
Thus it now requires viziphant to be installed or in $PYTHONPATH in
order to be able to import it.